lstm updates + hsfiles + spotpy calibration + teehr
